[+] splitted functions in multiple files, still need to fix some issues like coding style and case where file searched doesnt exist
