+++ src/formatting/globbings/command_parsing.c
@@ -1,0 +1,19 @@
+/*
+** EPITECH PROJECT, 2025
+** src/formatting/globbings/command_parsing
+** File description:
+** file where the parsing for autocompletion is done
+*/
+
+#include "core/minishell.h"
+
+// char **find_files_starting_with(char *path)
+// {
+//     char **files_found = NULL;
+//     char *processed_path = NULL;
+//
+//     if (path == NULL)
+//         return NULL;
+//     files_found = count_files(path);
+//     return NULL;
+// }
